moved parsing routines in cache_files and a new io/tools.py file
